// src/types/CheckoutInterface.ts
export interface CheckoutCompanyDetailsInterface {
  ReturnChallanNumber: string | null;
  SaleChallanNumber: string | null;
  FocChallanNumber: string | null;
  CompanyFromPanNumber: string | null;
  CompanyToPanNumber: string | null;
  CompanyFrom: string;
  CompanyTo: string;
  CompanyFromAddress: string;
  CompanyToAddress: string;
  CompanyFromPhoneNumber: string;
  CompanyToPhoneNumber: string;
  CompanyFromId: string;
  CompanyToGUId: string;
}
// ...
export interface CheckoutBranchDetailsInterface {
  SaleChallanNumber: string;
  BranchFrom: string;
  BranchTo: string;
  BranchFromAddress: string;
  BranchToAddress: string;
  BranchFromPhoneNumber: string;
  BranchToPhoneNumber: string;
  BranchFromId: string;
  BranchToGUId: string;
  ChallanNumber: string;
}

export interface SelectedInventoryDetailInterface {
  InventoryId: string;
  ExpiryDate: string;
  PartNumber: string;
  BatchNumber: string;
  ContainerType: string;
  PackSize: string;
  Quantity: number;
}

export interface CheckoutInventoryDescriptionInterface {
  Description: string;
  Quantity: string;
  ShortName: string;
  SelectedInventories: SelectedInventoryDetailInterface[];
}